make thread_local_basic work
The problem is that the compiler-generated future holds the !Send type.
So, the future becomes !Send.
Event though it doesn't live until the first await point.
The reason is that there's not enough data structure inlining in the
compiler when it generates that future, I think.

Impl it manually, with enough struct inlining, so that the !Send type
isn't part of the future.
More code per op than I'd want, let's trim that down in the next commit.