avoid overhead of constructing debug messages that are not printed (#54)
I noticed that the neon pageserver unit tests were taking much longer
with tokio-epoll-uring than with std:fs. It was because many of the
tests used tokio::time:pause(), and then did a call to
tokio::time::timeout() with a long 1h timeout. With tokio-epoll-uring,
the loop in poller_impl_impl() woke up 36000 times, i.e. once for every
100 ms of virtual time, before the timeout was reached. The significance
of that varies greatly depend on debug vs release build; it's
particularly bad when I use nightly compiler with cranelift codegen
backend.

It would be better to avoid the wakeups every 100 ms altogether, if the
debug-level events are not enabled, but it scares me to have such
different behavior depending on a debug-level. It sounds like a recipe
for very hard-to-debug bugs. Or perhaps the debug dumping could be just
removed altogether? That would be good for saving battery life when the
system is idle.