refactor: Use an explicit WaitOnCompletion future
This is easier to reason about than the implicit future with
scopeguard. The scopeguard::defer! action to place the resources in
the slot is now done in the future's drop method.

This moves the handling of YIELD_TO_EXECUTOR_IF_READY_ON_FIRST_POLL
from the future itself into the caller. I initially tried to implement
that logic in the future's poll() function, but that required tracking
the state of whether we are still waiting for the IO completion or
just yielding, which was more complicated. This lost the regression
test for https://github.com/neondatabase/tokio-epoll-uring/issues/37,
because I couldn't find a good way to construct that with the new
future anymore. But now that the caller is doing the yielding, that
kind of a bug cannot happen anymore anyway.